backend/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
SQLAlchemy==2.0.19
python-dotenv==1.0.0
